+++ Project/Source/App/Coordinator/Main/MainCoordinator.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 final class MainCoordinator: baseCoordinator {
     
     // MARK: - Start
@@ -12,14 +10,6 @@ final class MainCoordinator: baseCoordinator {
     // MARK: - Navigate
     override func navigate(to step: GabozagoStep) {
         switch step {
-        case .mainIsRequired:
-            navigateToLogin()
-        case .searchIsRequired:
-            searchIsRequired()
-        case .likePostIsRequired:
-            likePostIsRequired()
-        case .profileIsRequired:
-            profileIsRequired()
         case .createPostIsRequired:
             createPostIsRequired()
         }
@@ -27,30 +17,6 @@ final class MainCoordinator: baseCoordinator {
 }
 // MARK: - extension
 private extension MainCoordinator {
-    func navigateToLogin() {
-        let vm = MainViewModel(coordinator: self)
-        let vc = MainViewController(viewModel: vm)
-        self.navigationController.setViewControllers([vc], animated: true)
-    }
-    
-    func searchIsRequired() {
-        let vm = SearchViewModel(coordinator: self)
-        let vc = SearchViewController(viewModel: vm)
-        self.navigationController.setViewControllers([vc], animated: true)
-    }
-    
-    func likePostIsRequired() {
-        let vm = LikePostViewModel(coordinator: self)
-        let vc = LikePostViewController(viewModel: vm)
-        self.navigationController.setViewControllers([vc], animated: true)
-    }
-    
-    func profileIsRequired() {
-        let vm = ProfileViewModel(coordinator: self)
-        let vc = ProfileViewController(viewModel: vm)
-        self.navigationController.setViewControllers([vc], animated: true)
-    }
-    
     func createPostIsRequired() {
         let vm = CreatePostViewModel(coordinator: self)
         let vc = CreatePostVC(viewModel: vm)
