+++ Project/Source/App/Coordinator/Step/GabozagoStep.swift
@@ -1,10 +1,4 @@
 enum GabozagoStep {
-    // Main
-    case mainIsRequired
-    case searchIsRequired
-    case likePostIsRequired
-    case profileIsRequired
-    
     // Post
     case createPostIsRequired
 }
