+++ Project/Source/App/Coordinator/Step/TabBarStep.swift
@@ -1,0 +1,7 @@
+enum TabBarStep {
+    // TabBar
+    case mainIsRequired
+    case searchIsRequired
+    case likePostIsRequired
+    case profileIsRequired
+}
